fix: explicit NUMERIC casts in taste compatibility ROUND() expressions
Three RPCs used implicit type conversion in ROUND() expressions:
- get_taste_compatibility
- get_similar_taste_users
- get_local_lists_for_homepage

The math worked today because rating_10 is DECIMAL(3,1), but a future
type change to float/double would break silently. Explicit ::NUMERIC
casts before ROUND make the intent clear and prevent silent drift.

Migration: supabase/migrations/2026-04-10-numeric-casts.sql
Run in Denis's Supabase SQL Editor.

+++ supabase/schema.sql
@@ -1219,9 +1219,9 @@ LANGUAGE SQL STABLE SET search_path = public AS $$
   )
   SELECT
     COUNT(*)::INT AS shared_dishes,
-    ROUND(AVG(ABS(rating_a - rating_b)), 1) AS avg_difference,
+    ROUND(AVG(ABS(rating_a - rating_b))::NUMERIC, 1) AS avg_difference,
     CASE
-      WHEN COUNT(*) >= 3 THEN ROUND(100 - (AVG(ABS(rating_a - rating_b)) / 9.0 * 100))::INT
+      WHEN COUNT(*) >= 3 THEN ROUND((100 - (AVG(ABS(rating_a - rating_b))::NUMERIC / 9.0 * 100)))::INT
       ELSE NULL
     END AS compatibility_pct
   FROM shared;
@@ -1241,7 +1241,7 @@ RETURNS TABLE (
 LANGUAGE SQL STABLE SET search_path = public AS $$
   WITH candidates AS (
     SELECT b.user_id AS other_id, COUNT(*)::INT AS shared,
-      ROUND(100 - (AVG(ABS(a.rating_10 - b.rating_10)) / 9.0 * 100))::INT AS compat
+      ROUND((100 - (AVG(ABS(a.rating_10 - b.rating_10))::NUMERIC / 9.0 * 100)))::INT AS compat
     FROM votes a
     JOIN votes b ON a.dish_id = b.dish_id AND b.user_id != p_user_id AND b.rating_10 IS NOT NULL
     WHERE a.user_id = p_user_id AND a.rating_10 IS NOT NULL
@@ -2511,7 +2511,7 @@ AS $$
     CASE
       WHEN p_viewer_id IS NOT NULL AND p_viewer_id != ll.user_id THEN (
         SELECT CASE
-          WHEN COUNT(*) >= 3 THEN ROUND(100 - (AVG(ABS(a.rating_10 - b.rating_10)) / 9.0 * 100))::INT
+          WHEN COUNT(*) >= 3 THEN ROUND((100 - (AVG(ABS(a.rating_10 - b.rating_10))::NUMERIC / 9.0 * 100)))::INT
           ELSE NULL
         END
         FROM votes a
